refactor: remove previous class for lunatic loops
