+++ eno-core/src/main/java/fr/insee/eno/core/processing/LunaticProcessing.java
@@ -36,7 +36,7 @@ public class LunaticProcessing {
         new LunaticAddGeneratingDate().apply(lunaticQuestionnaire);
         //
         new LunaticSortComponents(enoQuestionnaire).apply(lunaticQuestionnaire);
-        new LunaticLoopResolution2(enoQuestionnaire).apply(lunaticQuestionnaire);
+        new LunaticLoopResolution(enoQuestionnaire).apply(lunaticQuestionnaire);
 
         new LunaticAddBindingDependencies(lunaticCatalog, enoIndex).apply(lunaticQuestionnaire);
         new LunaticAddMissingVariables(parameters.isMissingVariables()).apply(lunaticQuestionnaire);
